+++ packages/backend/src/services/InstanceConfigurationService/InstanceConfigurationService.test.ts
@@ -72,7 +72,7 @@ const createMockService = (overrides: AnyType = {}) => {
     };
 
     const projectModel = {
-        getProjectUuids: jest.fn(),
+        getDefaultProjectUuids: jest.fn(),
         getWithSensitiveFields: jest.fn(),
         update: jest.fn(),
         ...overrides.projectModel,
@@ -129,7 +129,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                         .mockResolvedValue(['org-1', 'org-2']),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue(['project-1', 'project-2']),
                 },
@@ -170,7 +170,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue(['project-1', 'project-2']),
                 },
@@ -197,7 +197,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                 },
@@ -249,7 +249,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                 },
@@ -292,7 +292,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                 },
@@ -346,7 +346,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                     getWithSensitiveFields: jest
@@ -389,7 +389,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                     getWithSensitiveFields: jest
@@ -444,7 +444,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                     getWithSensitiveFields: jest
@@ -477,7 +477,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                     getWithSensitiveFields: jest
@@ -505,7 +505,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                 },
@@ -529,7 +529,7 @@ describe('InstanceConfigurationService.updateInstanceConfiguration', () => {
                     getOrgUuids: jest.fn().mockResolvedValue([mockOrgUuid]),
                 },
                 projectModel: {
-                    getProjectUuids: jest
+                    getDefaultProjectUuids: jest
                         .fn()
                         .mockResolvedValue([mockProjectUuid]),
                     getWithSensitiveFields: jest