+++ packages/common/src/types/table.ts
@@ -28,5 +28,5 @@ export type TableBase = {
     requiredAttributes?: Record<string, string | string[]>;
     groupDetails?: Record<string, GroupType>;
     defaultTimeDimension?: DefaultTimeDimension;
-    aiHint?: string;
+    aiHint?: string | string[];
 };